rename all files to .hpp and fix mac os build
