+++ include/robot2D/Core/WindowContext.hpp
@@ -21,7 +21,7 @@ source distribution.
 
 #pragma once
 
-#include <robot2D/Config.h>
+#include <robot2D/Config.hpp>
 
 namespace robot2D {
     struct ROBOT2D_EXPORT_API WindowContext {
