+++ include/robot2D/Extra/App.hpp
@@ -25,9 +25,11 @@ source distribution.
 #include <memory>
 #include <cassert>
 #include <string>
+#include <robot2D/Core/Window.hpp>
 
-#include "State.h"
-#include "IStateMachine.h"
+
+#include "State.hpp"
+#include "IStateMachine.hpp"
 
 namespace robot2D{
 
@@ -55,6 +57,7 @@ namespace robot2D{
         float lastTime;
         float deltaTime;
     };
+
 
     template<class T, typename ... Args>
     void App::register_state(const int id, Args&& ... args) {
