+++ include/robot2D/Graphics/Font.hpp
@@ -20,28 +20,37 @@ source distribution.
 *********************************************************************/
 
 #pragma once
-#include <unordered_map>
-#include <string>
-#include <memory>
 
-#include <robot2D/Config.h>
+#include <string>
+#include <map>
+
+#include "robot2D/Core/Vector2.hpp"
+
 
 namespace robot2D{
-    template<typename T, typename ID = std::string>
-    class ROBOT2D_EXPORT_API ResourceHandler{
-    public:
-        using Ptr = std::unique_ptr<T>;
-    public:
-        ResourceHandler();
-        ~ResourceHandler() = default;
 
-        template<typename ... Args>
-        bool loadFromFile(const ID& idx, Args&&... args);
-
-        const T& get(const ID& idx) const;
-    private:
-        std::unordered_map<ID, Ptr> m_resources;
+    struct character {
+        unsigned int textureID; // ID handle of the glyph texture
+        vec2i   size;      // size of glyph
+        vec2i   bearing;   // offset from baseline to left/top of glyph
+        unsigned int advance;   // horizontal offset to advance to next glyph
     };
 
-    #include "ResourceHandler.inl"
+
+    class Font{
+    public:
+        Font();
+        ~Font();
+
+        bool loadFromFile(const std::string& path, const unsigned int& char_size = 20);
+        const std::map<char, character>& get_chars()const;
+    private:
+        void cleanup();
+        void setup_cache();
+        void generate_texture(unsigned int& texture, void* glyph);
+    private:
+        void* m_library;
+        void* m_face;
+        std::map<char, character> m_characters;
+    };
 }
