+++ include/robot2D/Graphics/Shader.hpp
@@ -22,7 +22,7 @@ source distribution.
 #pragma once
 
 #include <string>
-#include "robot2D/Core/Vector2.h"
+#include "robot2D/Core/Vector2.hpp"
 
 namespace robot2D{
     enum class shaderType {
